+++ changelog.md
@@ -5,7 +5,7 @@
 ### Features
 
 * Add red, green, blue, red-green, red-blue, blue-green and luma variant color comparisons for dithering and outline filters
-* Add color palette
+* Add color palettes
 
 ### Technical